+++ Otros/Entidades/Entidad.cs
@@ -6,6 +6,6 @@ namespace Bot_Dofus_1._29._1.Otros.Entidades
     public interface Entidad : IDisposable
     {
         int id { get; set; }
-        int celda_id { get; set; }
+        short celda_id { get; set; }
     }
 }
